+++ lib/chore.dart
@@ -1,0 +1,7 @@
+class Chore {
+  final String text;
+
+  // add deadline date?
+
+  Chore(this.text);
+}
