+++ lib/main.dart
@@ -1,5 +1,9 @@
 import 'package:flutter/material.dart';
 
+import 'chore_item.dart';
+
+import 'chore.dart';
+
 void main() {
   runApp(const MyApp());
 }
@@ -188,78 +192,4 @@ class ListPage extends StatelessWidget {
     );
     */
   }
-}
-
-Widget _chore(BuildContext context, Chore chore) {
-  return Column(
-    children: [
-      Padding(
-        padding: EdgeInsets.all(3),
-        child: Row(
-          children: [
-            IconButton(
-                icon: Icon(Icons.check_box_outline_blank_outlined),
-                tooltip: 'Mark as done',
-                onPressed: () {}
-                //toggleBox();, // Does nothing right now
-                ),
-            TextButton(
-                child: Text(chore.text),
-                onPressed: () {}), // does nothing right now, will be able to
-            // edit the text later when clicked
-            Spacer(), // to push the delete button to rightmost end of the row
-            IconButton(
-              icon: Icon(Icons.delete),
-              tooltip: 'Delete',
-              onPressed: () {}, // does nothing right now
-            ),
-          ],
-        ),
-      ),
-    ],
-  );
-}
-
-class ChoreItem extends StatelessWidget {
-  final Chore chore;
-  ChoreItem(this.chore, {super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Padding(
-          padding: EdgeInsets.all(3),
-          child: Row(
-            children: [
-              IconButton(
-                  icon: Icon(Icons.check_box_outline_blank_outlined),
-                  tooltip: 'Mark as done',
-                  onPressed: () {}
-                  //toggleBox();, // Does nothing right now
-                  ),
-              TextButton(
-                  child: Text(chore.text),
-                  onPressed: () {}), // does nothing right now, will be able to
-              // edit the text later when clicked
-              Spacer(), // to push the delete button to rightmost end of the row
-              IconButton(
-                icon: Icon(Icons.delete),
-                tooltip: 'Delete',
-                onPressed: () {}, // does nothing right now
-              ),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class Chore {
-  final String text;
-
-  // add deadline date?
-
-  Chore(this.text);
 }
